# prometheus collection memory improvements