```
benchmark                  old ns/op     new ns/op     delta
BenchmarkBuildTags1-8      172           128           -25.58%
BenchmarkBuildTags2-8      191           219           +14.66%
BenchmarkBuildTags4-8      239           354           +48.12%
BenchmarkBuildTags8-8      341           622           +82.40%
BenchmarkBuildTags16-8     1852          1203          -35.04%

benchmark                  old allocs     new allocs     delta
BenchmarkBuildTags1-8      2              1              -50.00%
BenchmarkBuildTags2-8      2              1              -50.00%
BenchmarkBuildTags4-8      2              1              -50.00%
BenchmarkBuildTags8-8      2              1              -50.00%
BenchmarkBuildTags16-8     4              1              -75.00%

benchmark                  old bytes     new bytes     delta
BenchmarkBuildTags1-8      336           32            -90.48%
BenchmarkBuildTags2-8      336           48            -85.71%
BenchmarkBuildTags4-8      336           80            -76.19%
BenchmarkBuildTags8-8      336           144           -57.14%
BenchmarkBuildTags16-8     2149          288           -86.60%
```